cleaning main.dart
